functions/run_job/requirements.txt
functions-framework==3.5.0
google-cloud-run==0.10.5
google-events==0.12.0
google-cloud-logging==3.10.0
ruamel.yaml==0.18.6
